requirements.txt
asgiref==3.10.0
black==25.9.0
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
distlib==0.4.0
Django==5.2.7
django-types==0.22.0
djangorestframework==3.16.1
djangorestframework-types==0.9.0
djangorestframework-simplejwt
filelock==3.20.0
identify==2.6.15
idna==3.11
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pre_commit==4.3.0
psycopg2-binary==2.9.11
pyright==1.1.407
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
requests==2.32.5
ruff==0.14.3
sqlparse==0.5.3
types-psycopg2==2.9.21.20251012
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.4
